fix: unblock the release gate — private path in P021 validation output, lint, one over-tight tolerance
The repository-wide path check failed on
products/P021/validation/keepout_cross_check_output.txt, which recorded the
build machine's absolute path for the KeepOut module. That single line blocked
publication of every product in the batch. The validation script now prints the
path relative to the repository root, and the output is regenerated.

P021 SlewForge: test_ninety_degree_quaternion asserted an absolute tolerance of
1e-16 on a value near 0.707, where one ULP is 1.1e-16 — below representable
resolution, so sin(pi/4) rounding one ULP low failed a correct implementation.
Tolerance widened to 1e-15; the comment in the test already documented the
expected one-ULP offset.

Lint: ruff clean across P021 and P022 (RUF022 __all__ ordering, I001 import
order, RUF046 redundant int() around round(), RUF059 unused SVD unpack, RUF007
zip-successive-pairs, ISC004 implicit concatenation). No behavioural change.

Gate after this change: P022 CMGSteer PASS, 285 tests. P021 SlewForge 32 tests
passing, lint clean.

+++ products/P022/src/cmgsteer/simulate.py
@@ -107,7 +107,7 @@ def constant_profile(
         raise ValueError(f"torque must have shape (3,), got {t.shape}")
     if duration <= 0.0 or dt <= 0.0:
         raise ValueError("duration and dt must be positive [s]")
-    n = max(1, int(round(duration / dt)))
+    n = max(1, round(duration / dt))
     return TorqueProfile(np.tile(t, (n, 1)), dt, name)
 
 
@@ -153,7 +153,7 @@ def rest_to_rest_profile(
         raise ValueError(f"momentum_change must be positive [N*m*s], got {momentum_change}")
     if duration <= 0.0 or dt <= 0.0:
         raise ValueError("duration and dt must be positive [s]")
-    n = max(2, int(round(duration / dt)))
+    n = max(2, round(duration / dt))
     t_mid = (np.arange(n) + 0.5) * dt
     if shape == "bang-bang":
         tau_max = 2.0 * momentum_change / duration
